feat: Actualizar navegación correcta entre ConsultarFAQ y RegistarFAQ
