+++ app/src/main/java/com/kotlin/sacalabici/framework/views/fragments/RegisterFAQFragment.kt
@@ -5,6 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.kotlin.sacalabici.R
 import com.kotlin.sacalabici.databinding.FragmentRegisterFaqBinding
 
 class RegisterFAQFragment : Fragment() {
@@ -19,6 +20,9 @@ class RegisterFAQFragment : Fragment() {
         // Inflar el layout y obtener el binding
         _binding = FragmentRegisterFaqBinding.inflate(inflater, container, false)
 
+        // Botón para abrir RegisterFAQFragment
+        returnToFAQ()
+
         // Retorna la vista raíz del binding
         return binding.root
     }
@@ -26,5 +30,18 @@ class RegisterFAQFragment : Fragment() {
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
+    }
+
+    // Función para regresar a FAQFragment
+    private fun returnToFAQ() {
+        val btnFAQs = binding.BCerrar
+        btnFAQs.setOnClickListener {
+            // Navegar a RegisterFAQsFragment y reemplazar el contenido en el contenedor principal de MainActivity
+            parentFragmentManager
+                .beginTransaction()
+                .replace(R.id.nav_host_fragment_content_main, FAQFragment())
+                .addToBackStack(null) // Para permitir navegar hacia atrás
+                .commit()
+        }
     }
 }
